requirements.txt
torch>=2.0.0
transformers>=4.38.0
sentence-transformers>=2.6.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
datasets>=2.18.0       # HuggingFace datasets (HaluEval, STS-B, SNLI, CNN/DM)
nltk>=3.8.1
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.13.0
